backend/app/requirements.txt
fastapi
uvicorn[standard]
pandas
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
